requirements.txt
composio-core>=0.3.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
sqlalchemy>=2.0.0
